Made IntegrationTestCase into a trait
